net: shell: Add udp command
Move "udp" command to a separate file.
